Make OmniSharp launch settings more robust
The extension looks in three locations for OmniSharp:

1. Whatever the user specified in the "csharp.omnisharp" setting.
2. The value of the "OMNISHARP" environment variable.
3. The expected location where the bundled OmniSharp is copied.

If #1 was set, #2 and #3 would be ignored even if the value of #1 was invalid. Likewise, if #2 was set, #3 would be ignored even if the value of #2 was invalid. This change adds fallback logic to try the next option if an invalid value is encountered. In addition, it is more robust depending on whether a file or directory name is specified. If the value is a directory, it will try 'run' (run.cmd on Windows) or 'omnisharp' (omnisharp.exe on Windows) as possible launch file names.